src/backend-api/requirements.txt
# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# RAG Core
llama-index>=0.9.0
llama-index-readers-docling>=0.1.0
llama-index-vector-stores-postgres>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-llms-ollama>=0.1.0

# Native Reranking 
sentence-transformers>=2.2.2
transformers>=4.21.0
torch>=1.12.0
numpy>=1.21.0

# Agentic System
crewai>=0.32.0
crewai[tools]>=0.32.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Observability
arize-phoenix>=4.0.0

# Document Processing
docling>=1.0.0
python-docx>=0.8.11

# Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
# FastAPI Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# RAG Core
llama-index>=0.9.0
llama-index-readers-docling>=0.1.0
llama-index-vector-stores-postgres>=0.1.0
llama-index-embeddings-ollama>=0.1.0
llama-index-llms-ollama>=0.1.0
rank_bm25==0.2.2
# Native Reranking (Key Addition)
sentence-transformers>=2.2.2
transformers>=4.21.0
torch>=1.12.0
numpy>=1.21.0

# Agentic System
crewai>=0.32.0
crewai[tools]>=0.32.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Observability
arize-phoenix>=4.0.0
openinference-instrumentation>=0.1.27
# Document Processing
docling>=1.0.0
python-docx>=0.8.11

# Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0


